pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "altair[all]>=5.5.0",
    "geopandas>=1.0.1",
    "matplotlib>=3.10.1",
    "psycopg[binary]>=3.2.6",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
    "streamlit-folium>=0.25.0",
]
